// backend/src/common/messages/message.enum.ts
export enum MessageCode {
  INVALID_CREDENTIALS = 'INVALID_CREDENTIALS',
  TOKEN_NOT_PROVIDED = 'TOKEN_NOT_PROVIDED',
  TOKEN_EXPIRED = 'TOKEN_EXPIRED',
  INVALID_TOKEN = 'INVALID_TOKEN',
  USER_ALREADY_EXISTS = 'USER_ALREADY_EXISTS',
  VERIFY_EMAIL_SUBJECT = 'VERIFY_EMAIL_SUBJECT',
  VERIFY_EMAIL_TEMPLATE = 'VERIFY_EMAIL_TEMPLATE',
  CORS_ORIGIN_MISSING = 'CORS_ORIGIN_MISSING',
  USER_NOT_FOUND = 'USER_NOT_FOUND',
  RESET_PASSWORD_SUBJECT = 'RESET_PASSWORD_SUBJECT',
  RESET_EMAIL_TEMPLATE = 'RESET_EMAIL_TEMPLATE',
  UNAUTHORIZED = 'UNAUTHORIZED',
  FORBIDDEN = 'FORBIDDEN',
  BAD_REQUEST = 'BAD_REQUEST',
  INTERNAL_SERVER = 'INTERNAL_SERVER',
  REFRESH_TOKEN_IS_USED = 'REFRESH_TOKEN_IS_USED',
  USER_NOT_ACTIVE = 'USER_NOT_ACTIVE',
  OVER_MAX_ATTEMPT_LOGIN = 'OVER_MAX_ATTEMPT_LOGIN',
  INVALID_ROOM_CODE = 'INVALID_ROOM_CODE',
  ROOM_NOT_FOUND = 'ROOM_NOT_FOUND',
  USER_ALREADY_IN_ROOM = 'USER_ALREADY_IN_ROOM',
  USER_NOT_IN_ROOM_MEMBER = 'USER_NOT_IN_ROOM_MEMBER'
}